Enable reusing ccache directories.
Commit 2d6d6ba119e2bf303f82dbc3ce07e8120e7e3914 enabled building with
ccache, but there was no mechanism to save and restore caches. This
commit adds two functions, download_ccache and upload_ccache, which
download and upload a tarred ccache directory.

The cache is uploaded to s3 at the end of a full build if
UPLOAD_CCACHE=1.

The 3 variables that affect the handling of reusing ccache are:
DOWNLOAD_CCACHE = Downloads ccache from s3.
USE_CCACHE = Modifies environment variables so that ccache is called.
UPLOAD_CCACHE = Uploads a ccache tarball to s3.

DOWNLOAD_CCACHE and USE_CCACHE now default to 1.

Change-Id: I482aa13e833d4680efe7cab98aad7f4fb998bfc0